+++ src/sumologiccse/__init__.py
@@ -13,7 +13,7 @@ from .sumologiccse import (
     SumoLogicCSEError,
 )
 
-__version__ = "0.1.0"
+__version__ = "0.2.0"
 __all__ = [
     "SumoLogicCSE",
     "SumoLogicCSEError",